ディレクトリの変更とstylecomponentの変更
